- exclusive users flow done - product and merchant page done - upload to play store (exclusive user 0.1 version)

+++ app/src/main/java/com/bejohen/pikapp/view/home/HomeCategoryListAdapter.kt
@@ -35,7 +35,7 @@ class HomeCategoryListAdapter(val categoryItemlist: ArrayList<ItemHomeCategory>)
     }
 
     override fun onCategoryClicked(v: View) {
-        val uuid = v.HomeCategoryUuid.text.toString().toInt()
+        val uuid = v.HomeCategoryUuid.text.toString().toLong()
         val action = HomeFragmentDirections.actionToCategoryFragment(uuid)
         Navigation.findNavController(v).navigate(action)
     }
